Framework for Social Network using present access policy
This paper is about access policies using present technologies in social
networks which had become part of our life and social networking . There
various access control models for social networking systems, but not
cent percent suitable for it. The proposed framework for access control
model can bring the revolution in the field of Social Networking System
for Web 2.0 and it can be implemented
using presently available access control model.